Move ProvGraph into ProvVis, and retitle accordingly
