+++ context/app/static/js/components/ProvGraph/__tests__/fixtures/complex-cwl.json
