+++ context/app/static/js/components/ProvGraph/__tests__/fixtures/real-cwl.json
