+++ context/app/static/js/components/ProvGraph/__tests__/fixtures/simple-cwl.json
